+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 1.13.0
 
-* Add a `collectBytes` function which collects list-of-byte events into
-  a single byte list.
+* Add `collectBytes` and `collectBytesCancelable` functions which collects
+  list-of-byte events into a single byte list.
 
 * Fix a bug where rejecting a `StreamQueueTransaction` would throw a
   `StateError` if `StreamQueue.rest` had been called on one of its child queues.